+++ backend/footprint/src/main/java/foot/footprint/domain/article/application/edit/EditArticleService.java
@@ -1,0 +1,6 @@
+package foot.footprint.domain.article.application.edit;
+
+public interface EditArticleService {
+
+    public void edit(Long articleId, Long memberId, String newContent);
+}
